requirements.txt
pkg-resources==0.0.0
python-ptrace==0.9.5
six==1.15.0
sysv-ipc==1.0.1
